+++ CHANGELOG.md
@@ -18,6 +18,7 @@ changes.
 
 - Fix calculating votes counting for governance actions
 - Fix crashing backend on unhandled missing proposal from vote [Issue 2920](https://github.com/IntersectMBO/govtool/issues/2920)
+- Remove abstain votes (not auto abstain) from total DRep stake
 
 ### Changed